+++ config/packages/csrf.yaml
@@ -1,11 +1,6 @@
 # Enable stateless CSRF protection for forms and logins/logouts
 framework:
-    form:
-        csrf_protection:
-            token_id: submit
+    # ...
+    csrf_protection: ~
 
-    csrf_protection:
-        stateless_token_ids:
-            - submit
-            - authenticate
-            - logout
+
