+++ src/Controller/UsersController.php
@@ -52,25 +52,5 @@ final class UsersController extends AbstractController
         return $this->redirectToRoute('app_users_index', [], Response::HTTP_SEE_OTHER);
     }
 
-    #[Route('/profile/{id}', name: 'app_user_profile', methods: ['GET'])]
-    public function profile(EntityManagerInterface $entityManager) : Response
-    {
-        $user = $this->getUser();
-        if (!$user) {
-            
-            return $this->redirectToRoute('app_users_index', [], Response::HTTP_SEE_OTHER);
-        }
-    
-
-        
-
-        $tweets = $entityManager
-        ->getRepository(Tweets::class)
-        ->findBy(['user' => $user], ['createdAt' => ['DESC']]);
-
-        return $this->render('users/profile.html.twig',[
-            'user' =>$user,
-            'tweets' => $tweets,
-        ]);
-    }
+  
 }
